Finally understand the abrupt worker test failure
There are two identifiers associated with workers when working with an
overlay manager.

1. The "pid" (process ID) assigned to the worker by the ClusterManager
   internals
2. The "oid" (overlay ID) assigned to the worker by the overlay manager

Ideally these values would be the same but the "pid" is only available
once the worker and manager have established a connection with each
other. The workers need to have an overlay identifier when they are
launched in order receive messages on the overlay network. In order to
solve this we generate oid's at the worker spawn time.

The problem with the abrupt worker test failure is confusion between
the pid/oid. I am killing the worker with the oid 2 but that worker may
have the pid 3 if it happens to spawn after the other worker. The
modifications to the test demostrate this.