Remove .git directory in Docker image
When Pkg.update works with git directories if the HEAD is a branch it
expects that branch to be tracked (be associated with a remote branch).
In some scenarios there isn't a valid remote which is problematic:

```
fatal: '/home/gitlab-runner/builds/494e880d/0/invenia/AWSClusterManagers.jl' does not appear to be a git repository
fatal: Could not read from remote repository.
```

If the Julia package isn't a git repository we ignore all these
complications. The one down side is there isn't a good way to determine
what code revision is being used from inside the Docker image. However
since we're tagging these images using the SHA I think this isn't a big
problem at the moment.