Reduce Docker build time during development
When developing this package then testing it the Dockerfile would
be required to re-build the package requirements. With the current
changes the package requirements only need to be rebuilt if they have
changed.

As a consequence of this developers may need to remove older Docker
images or use manually build with `--no-cache` to force the requirements
to be rebuild and possible updated.