Refactor container launching
We don't need to have a separate task listening for each workers
connection as if multiple workers connect at the same time the worker
will wait until the server responds to its connection.

It is possible that with multiple concurrent tasks accepting connections
that connections could possibly be lost? If this is true this could be
one reason AWS Batch worker jobs fail to launch.