Drop support for running tests with Docker
Note that Docker can still be used as a worker but the Docker container
being built just no longer includes the "test" directory for this
package. Mainly this support was dropped as changing the tests would
result in the Docker image needing to be re-built and pushed which
slowed down iteration when developing new tests.

I did look into supporting this feature by optionally copying the test
directory based upon a --build-arg however it seems like there is no
way to do a conditional COPY in this way which would allow for the speed
benefits of building to exist.