Fix worker termination during cluster setup
Due to wanting to set the order of workers to match the node index
ordering we end up keeping workers alive for a period of time before the
manager will connect to them. If we wait longer than the default worker
timeout then worker which have waited longer than 60 seconds may
terminate themselves while the cluster is being launched and the cluster
manager will become stuck.

This fix increases the default worker timeout to 2 minutes or the same
amount of time we have the manager wait for workers to report in.