Fix missing NAT in us-east-1d
When the Compute Environment would launch machines in the us-east-1d
region those machines were unable to connect to the ECS endpoint as they
were launched in a private subnet which did not have a NAT associated.
This would result in AWS Batch jobs becoming stuck in the RUNNABLE
status as the Compute Environment would not launch additional resources
as it knew an instance had been launched.